requirements.txt
-e .
aws-cdk.aws_lambda_python
aws-cdk.aws_apigateway
aws-cdk.aws_route53_targets
python-dotenv